ml-service/requirements.txt
Flask==2.3.2
pandas==1.5.3
scikit-learn==1.2.2
numpy==1.24.3
python-dotenv==1.0.0
Werkzeug==2.3.6
